pyproject.toml
[project]
name = "raggr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "python-dotenv>=1.0.0",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "ollama>=0.6.0",
    "openai>=2.0.1",
    "pydantic>=2.11.9",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "black>=25.9.0",
    "pillow-heif>=1.1.1",
]
